#456921 [C#テキスト]4章演習問題
AddOneMonthメソッドを問題文に合わせ、新しくインスタンスを作る形に変更しました。
