+++ 章末問題4/Practice4-1-4/Program4-1-4.cs
@@ -3,7 +3,7 @@ namespace Practice4_1_4 {
     class Program {
         //Practice4-1-4 ToStringメソッドをオーバーライドしてください。結果は"2017年8月"といった形式にしてください。
         static void Main(string[] args) {
-            YearMonth wYearMonth = new YearMonth (2000, 3);
+            var wYearMonth = new YearMonth (2000, 3);
             Console.WriteLine(wYearMonth);
         }
     }
