+++ 章末問題4/Practice4-2-3/YearMonth.cs
@@ -21,7 +21,7 @@
             set { Is21stCentury = value; }
         }
         /// <summary>
-        /// YearMonthのYearMonthのコンストラクタ
+        /// YearMonthのコンストラクタ
         /// </summary>
         /// <param name="vYear">年</param>
         /// <param name="vMonth">月</param>
@@ -33,15 +33,34 @@
         /// 1か月後を求めるメソッド
         /// </summary>
         /// <returns>1か月後の年と日</returns>
-        public void AddOneMonth() {
-            this.Month++;
-            if (this.Month > 12) {
-                this.Month = 1;
-                this.Year++;
+        public YearMonth AddOneMonth() {
+            int wNewYear = this.Year;
+            int wNewMonth = this.Month + 1;
+            if (wNewMonth > 12) {
+                wNewMonth = 1;
+                wNewYear++;
             }
+            return new YearMonth(wNewYear, wNewMonth);
         }
+        /// <summary>
+        /// 書式を"2017年8月"の形に直すメソッド
+        /// </summary>
+        /// <returns>指定書式に変更した年月</returns>
         public override string ToString() {
             return ($"{Year}年{Month}月 ");
+        }
+        /// <summary>
+        /// 21世紀の年月を探すメソッド
+        /// </summary>
+        /// <param name="wYearMonths">年月</param>
+        /// <returns>21世紀の年月</returns>
+        public static YearMonth FindFirst21YearMonth(YearMonth[] wYearMonths) {
+            foreach (YearMonth wYearMonth in wYearMonths) {
+                if (wYearMonth.Is21stCentury) {
+                    return wYearMonth;
+                }
+            }
+            return null;
         }
     }
 }
