+++ 章末問題4/Practice4-2-5/Program4-2-5.cs
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 
 namespace Practice4_2_5 {
     class Program {
-        //Practice4-2-5 配列に格納されているすべてのYearMonthの1ヶ月後を求め、その結果を新たな配列に入れてください。その後、その配列の要素の内容（年月）を順に表示してください。LINQを使えると頃はLINQを使って実装してみてください。
+        //Practice4-2-5 配列に格納されているすべてのYearMonthの1ヶ月後を求め、その結果を新たな配列に入れてください。その後、その配列の要素の内容（年月）を順に表示してください。LINQを使える所はLINQを使って実装してみてください。
         static void Main(string[] args) {
              var wDay1 = new YearMonth(2000, 3);
             var wDay2 = new YearMonth(2100, 12);
@@ -14,7 +14,7 @@ namespace Practice4_2_5 {
             var wDay4 = new YearMonth(2003, 10);
             var wDay5 = new YearMonth(2450, 7);
             var wYearMonths = new YearMonth[] { wDay1, wDay2, wDay3, wDay4, wDay5, };
-            var wFirst21YearMonth = FindFirst21YearMonth(wYearMonths);
+            var wFirst21YearMonth = YearMonth.FindFirst21YearMonth(wYearMonths);
             for (int i = 0; i < wYearMonths.Length; i++){
                 wYearMonths[i].AddOneMonth();
             }
@@ -23,19 +23,6 @@ namespace Practice4_2_5 {
                 Console.WriteLine(wYearMonth.ToString());
             }
             Console.WriteLine($"最初に見つかった21世紀の年: {(wFirst21YearMonth != null ? wFirst21YearMonth.Year.ToString() + "年" : "見つかりませんでした")}");
-        }
-        /// <summary>
-        /// 21世紀の年月を探すメソッド
-        /// </summary>
-        /// <param name="wYearMonths">年月</param>
-        /// <returns>21世紀の年月</returns>
-        public static YearMonth FindFirst21YearMonth(YearMonth[] wYearMonths) {
-            foreach (YearMonth wYearMonth in wYearMonths) {
-                if (wYearMonth.Is21stCentury) {
-                    return wYearMonth;
-                }
-            }
-            return null;
         }
     }
